Move column conditional tests to postgres dir
  * Using postgres directory once only postgres will be compatible
  with conditional column operations.
